API: move TrackCurrentDoc and RenamePrivateArguments to sphinx.util
