+++ sphinx/base/conf_base.py
@@ -8,7 +8,7 @@ import logging
 import os
 import shutil
 import sys
-from typing import Any, Dict, Optional, Tuple, Type, cast
+from typing import Any, Dict, Optional, Tuple
 
 from sphinx.application import Sphinx
 
@@ -192,122 +192,16 @@ def setup(app: Sphinx) -> None:
     :param app: the Sphinx application object
     """
 
-    from types import FunctionType
-
-    from pytools.meta import SingletonABCMeta
-    from pytools.sphinx import AutodocBeforeProcessSignature, AutodocProcessSignature
     from pytools.sphinx.util import (
         AddInheritance,
         CollapseModulePathsInDocstring,
         CollapseModulePathsInSignature,
         CollapseModulePathsInXRef,
+        RenamePrivateArguments,
         Replace3rdPartyDoc,
-        ResolveTypeVariables,
         SkipIndirectImports,
+        TrackCurrentDoc,
     )
-
-    # todo: move this class to package pytools.sphinx.util once we release pytools 2.1
-    class TrackCurrentDoc(
-        AutodocProcessSignature,  # type: ignore
-        metaclass=SingletonABCMeta,
-    ):
-        """
-        Keep track of the class currently being processed by autodoc.
-
-        This is required to attribute unbound methods to the correct class, e.g.,
-        in class :class:`.ResolveTypeVariables`.
-        """
-
-        def __init__(self) -> None:
-            self.current_class: Optional[Type[Any]] = None
-            self.rtv = ResolveTypeVariables()
-
-        def process(
-            self,
-            app: Sphinx,
-            what: str,
-            name: str,
-            obj: object,
-            options: object,
-            signature: Optional[str],
-            return_annotation: Optional[str],
-        ) -> Optional[Tuple[Optional[str], Optional[str]]]:
-            if what == "class":
-                cls = cast(type, obj)
-                self.current_class = cls
-                self.rtv._update_current_class(cls)
-
-            return None
-
-        def connect(self, app: Sphinx, priority: Optional[int] = None) -> int:
-            self.rtv.connect(app, priority)
-            return cast(int, super().connect(app, priority))
-
-    class RenamePrivateArguments(
-        AutodocBeforeProcessSignature,  # type: ignore
-        metaclass=SingletonABCMeta,
-    ):
-        """
-        Rename private argument names to their original names given in the source code.
-
-        For example, arg ``__x`` of method ``f`` in
-
-        .. code-block:: python
-
-            class A:
-                def f(self, __x: int) -> None:
-                    ...
-
-        will be renamed from its internal, private name ``_A__x`` back to ``__x``.
-        """
-
-        def process(self, app: Sphinx, obj: Any, bound_method: bool) -> None:
-            if not (bound_method or isinstance(obj, FunctionType)):
-                return
-
-            # get the name of the module or class that the function is a member of
-            try:
-                containers = obj.__qualname__.split(".")
-            except AttributeError:
-                return
-
-            if len(containers) < 2:
-                return
-
-            private_prefix = f"_{containers[-2]}__"
-
-            def get_public_name(name: str) -> str:
-                if name.startswith(private_prefix):
-                    return name[len(private_prefix) - 2 :]
-                else:
-                    return name
-
-            # get the original signature
-            try:
-                annotations: Dict[str, Any] = obj.__annotations__
-            except AttributeError:
-                annotations = {}
-
-            annotations_original = list(annotations.items())
-            for name, annotation in annotations_original:
-                public_name = get_public_name(name)
-                if public_name is not name:
-                    del annotations[name]
-                    annotations[public_name] = annotation
-
-            try:
-                code = obj.__code__
-                arg_and_variable_names = code.co_varnames
-                if any(
-                    name.startswith(private_prefix) for name in arg_and_variable_names
-                ):
-                    obj.__code__ = code.replace(
-                        co_varnames=tuple(
-                            get_public_name(name) for name in arg_and_variable_names
-                        ),
-                    )
-            except AttributeError:
-                pass
 
     AddInheritance(collapsible_submodules=intersphinx_collapsible_submodules).connect(
         app=app
