+++ src/pytools/sphinx/util/_util.py
@@ -35,6 +35,7 @@ from typing import (
 import typing_inspect
 
 from ...api import AllTracker, get_generic_bases, inheritdoc, public_module_prefix
+from ...meta import SingletonABCMeta
 from .. import (
     AutodocBeforeProcessSignature,
     AutodocProcessDocstring,
@@ -96,9 +97,11 @@ __all__ = [
     "CollapseModulePathsInDocstring",
     "CollapseModulePathsInSignature",
     "CollapseModulePathsInXRef",
+    "RenamePrivateArguments",
     "Replace3rdPartyDoc",
     "ResolveTypeVariables",
     "SkipIndirectImports",
+    "TrackCurrentDoc",
 ]
 
 #
@@ -1023,6 +1026,102 @@ class ResolveTypeVariables(AutodocBeforeProcessSignature):
             self._resolve_attribute_signatures(cls=cls)
 
         return bindings
+
+
+class TrackCurrentDoc(AutodocProcessSignature, metaclass=SingletonABCMeta):
+    """
+    Keep track of the class currently being processed by autodoc.
+
+    This is required to attribute unbound methods to the correct class, e.g.,
+    in class :class:`.ResolveTypeVariables`.
+    """
+
+    def __init__(self) -> None:
+        self.current_class: Optional[Type[Any]] = None
+        self.rtv = ResolveTypeVariables()
+
+    def process(
+        self,
+        app: Sphinx,
+        what: str,
+        name: str,
+        obj: object,
+        options: object,
+        signature: Optional[str],
+        return_annotation: Optional[str],
+    ) -> Optional[Tuple[Optional[str], Optional[str]]]:
+        if what == "class":
+            cls = cast(type, obj)
+            self.current_class = cls
+            self.rtv._update_current_class(cls)
+
+        return None
+
+    def connect(self, app: Sphinx, priority: Optional[int] = None) -> int:
+        self.rtv.connect(app, priority)
+        return super().connect(app, priority)
+
+
+class RenamePrivateArguments(AutodocBeforeProcessSignature, metaclass=SingletonABCMeta):
+    """
+    Rename private argument names to their original names given in the source code.
+
+    For example, arg ``__x`` of method ``f`` in
+
+    .. code-block:: python
+
+        class A:
+            def f(self, __x: int) -> None:
+                ...
+
+    will be renamed from its internal, private name ``_A__x`` back to ``__x``.
+    """
+
+    def process(self, app: Sphinx, obj: Any, bound_method: bool) -> None:
+        if not (bound_method or isinstance(obj, FunctionType)):
+            return
+
+        # get the name of the module or class that the function is a member of
+        try:
+            containers = obj.__qualname__.split(".")
+        except AttributeError:
+            return
+
+        if len(containers) < 2:
+            return
+
+        private_prefix = f"_{containers[-2]}__"
+
+        def get_public_name(name: str) -> str:
+            if name.startswith(private_prefix):
+                return name[len(private_prefix) - 2 :]
+            else:
+                return name
+
+        # get the original signature
+        try:
+            annotations: Dict[str, Any] = obj.__annotations__
+        except AttributeError:
+            annotations = {}
+
+        annotations_original = list(annotations.items())
+        for name, annotation in annotations_original:
+            public_name = get_public_name(name)
+            if public_name is not name:
+                del annotations[name]
+                annotations[public_name] = annotation
+
+        try:
+            code = obj.__code__
+            arg_and_variable_names = code.co_varnames
+            if any(name.startswith(private_prefix) for name in arg_and_variable_names):
+                obj.__code__ = code.replace(
+                    co_varnames=tuple(
+                        get_public_name(name) for name in arg_and_variable_names
+                    ),
+                )
+        except AttributeError:
+            pass
 
 
 #
